% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atm_vap.R
\name{atm_vap}
\alias{atm_vap}
\title{Atmospheric vapor pressure}
\usage{
atm_vap(W, es)
}
\arguments{
\item{es}{Saturation vapor pressure, \eqn{e_s} (\eqn{kPa}).}
}
\description{
\code{atm_vap} returns the ambient atmospheric vapor pressure, \eqn{e_{a}}
(\eqn{kPa}), given a relative humidity and saturation vapor pressure for a
particular temperature.
}
\details{
Atmospheric vapor pressure (Dingman 2008):

\deqn{e_{a} = W \cdot e_{s}}

Dingman SL. 2008. Physical hydrology. Waveland Press.
}
\examples{

}
