% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{decimal_to_permille}
\alias{decimal_to_permille}
\title{Decimal to permille converstion}
\usage{
decimal_to_permille(delta)
}
\arguments{
\item{delta}{A \eqn{\delta} value in decimal units.}
}
\value{
Numeric value.
}
\description{
A convenience function that converts a decimal delta value to units of
permillle.
}
