% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{delta_to_ratio}
\alias{delta_to_ratio}
\title{Convert delta value to isotope ratio}
\usage{
delta_to_ratio(delta, element, standard = "VSMOW", unit)
}
\arguments{
\item{delta}{An isotope delta value, in per mille.}

\item{element}{Character value indicating either "Hydrogen" or "Oxygen".}

\item{standard}{Character value indicating the acronym of the standard to be
used: "VMSOW", "GISP", or "SLAP".}

\item{unit}{Character value explicitly indicating if the input units are
"decimal" or "permille".}
}
\value{
Unitless numeric value of the isotope ratio.
}
\description{
This function, \code{delta_to_ratio} takes a delta value, in per mille of
either hydrogen or oxygen, and converts it to an isotope ratio [-], using a
standard. Applicable standards include Vienna Standard Mean Ocean Water
(VSMOW), Greenland Ice Sheet Precipitation (GISP), and Standard Light
Antarctic Precipitation (SLAP).
}
\details{
The relationship used to convert a given delta value to an isotope ratio:

\deqn{{}^{A}R = \Bigg(\frac{\delta^{A}X}{1000} + 1\Bigg) {}^{A}R_{Standard}}

Where \eqn{A} indicates the isotope of element \eqn{X}, \eqn{R} is the ratio
of the sample, and \eqn{R_{Standard}} is the isotope ratio of the standard
being used.

Below are the values of each standard and its isotope ratio used in this
function.

\tabular{lrr}{
Standard \tab 2H/1H \tab 18O/16O\cr
VSMOW \tab 0.00015576 \tab 0.00200520\cr
GISP \tab 0.00012624 \tab 0.00195555\cr
SLAP \tab 0.00008902 \tab 0.00189391 }

VSMOw and SLAP standard information can be found at:
https://nucleus.iaea.org/rpst/documents/VSMOW_SLAP.pdf

GISP standard information:
https://nucleus.iaea.org/rpst/Documents/rs_GISP.pdf
}
