% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E_Hamon.R
\name{E_Hamon}
\alias{E_Hamon}
\title{Lake evaporation (Hamon method)}
\usage{
E_Hamon(D, ps, a = 0.55)
}
\arguments{
\item{D}{Hours of daylight, \eqn{D} (\eqn{hrs}).}

\item{ps}{saturation vapor density, \eqn{\rho_{s}} (\eqn{g m^{-3}}).}

\item{a}{An empirical coefficient, assumed to be 0.55.}
}
\description{
\code{E_Hamon} provides an estimate of daily evaporation, \eqn{E} (\eqn{mm
day^{-1}}) using data related to day length.
}
\details{
Daily evaporation is determined via (Hamon 1961, Rosenberry et al. 2007):

\deqn{a \bigg(\frac{D}{12}\bigg)^{2} \frac{\rho_{s}}{100} \cdot 25.4}

Hamon, WR. 1961. Estimating potential evapotranspiration. Proc. Amer. Soc.
civ. Engrs. 87, 107–120.

Rosenberry DO, Winter TC, Buso DC, Likens GE. 2007. Comparison of 15
evaporation methods applied to a small mountain lake in the northeastern USA.
Journal of Hydrology 340 (3–4): 149–166. DOI: 10.1016/j.jhydrol.2007.03.018.
}
\examples{

}
