% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E_MassTransfer.R
\name{E_MassTransfer}
\alias{E_MassTransfer}
\title{Lake evaporation (Mass Transfer method)}
\usage{
E_MassTransfer(U, els, ea, N)
}
\arguments{
\item{U}{Windspeed 2 m above the lake surface, \eqn{U} (\eqn{m s^{-1}}).}

\item{els}{Saturated vapor pressure at the lake surface, \eqn{e_{ls}}
(\eqn{kPa}).}

\item{ea}{Atmospheric vapor pressure, \eqn{e_{a}} (\eqn{kPa}).}

\item{N}{Empirical parameter, no standard value.}
}
\description{
\code{E_MassTransfer} estimates a daily rate of evaporation, \eqn{E}
(\eqn{mm/day}), using a Dalton-type approach.
}
\details{
Evaporation as determined by (Harbeck et al. 1958, Rosenberry et al. 2007):

\deqn{E = (N U (e_{ls} - e_{a})) \cdot 10}

Harbeck GE, Kohler MA, Koberg GE. 1958. Water-loss investigations: Lake Mead
studies Available at: http://pubs.er.usgs.gov/publication/pp298.

Rosenberry DO, Winter TC, Buso DC, Likens GE. 2007. Comparison of 15
evaporation methods applied to a small mountain lake in the northeastern USA.
Journal of Hydrology 340 (3–4): 149–166. DOI: 10.1016/j.jhydrol.2007.03.018.
}
\examples{

}
