% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E_Papadakis.R
\name{E_Papadakis}
\alias{E_Papadakis}
\title{Lake evaporation (Papadakis method)}
\usage{
E_Papadakis(esmax, esmin, day, a = 0.5625, b = 2, conv = 10)
}
\arguments{
\item{esmax}{Maximum saturation vapor pressure for a given day,
\eqn{e_{s,max}} (\eqn{kPa}).}

\item{esmin}{Minimum saturation vapor pressure for a given day,
\eqn{e_{s,min}} (\eqn{kPa}).}

\item{day}{Number of days in a month, \eqn{day} (\eqn{days}).}

\item{a}{An emperical coefficient, assumed to be 0.5625.}

\item{b}{An emperical coefficient, assumed to be 2.}

\item{conv}{A multiplier that converts base units to mm/day. It is assumed to
be 10, but will need to be adjust for alternative units.}
}
\description{
\code{E_Papadakis} returns an estimate of evaporation, \eqn{E} (\eqn{mm
day^{-1}}), based on data related to saturation vapor pressure and number of
days in a month.
}
\details{
\eqn{E} is determined by (Papadakis 1965, McGuinness and Bordne 1972,
Rosenberry et al. 2007):

\deqn{E = a (e_{s,max} - [e_{s,min} - b]) \bigg(\frac{10}{day} \bigg)}

Papadakis J. 1965. Potential evapotranspiration. Bueuos Aires, 54 pp.

McGuinness JL, Bordne EF. 1972. A comparison of lysimeter-derived potential
evapotranspiration with computed values. Technical Bulletin 1452, US
Department of Agriculture Agricultural Research Service, Washington, DC.

Rosenberry DO, Winter TC, Buso DC, Likens GE. 2007. Comparison of 15
evaporation methods applied to a small mountain lake in the northeastern USA.
Journal of Hydrology 340 (3–4): 149–166. DOI: 10.1016/j.jhydrol.2007.03.018.
}
\examples{

}
