% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E_Thornthwaite.R
\name{E_Thornthwaite}
\alias{E_Thornthwaite}
\title{Lake evaporation (Thornethwaite method)}
\usage{
E_Thornthwaite(Ta, I, day, a = 1.6, b = 6.75e-07, c = 7.71e-05,
  d = 0.0179, e = 0.49)
}
\arguments{
\item{Ta}{Air temperature, \eqn{T_{a}} (\eqn{C}).}

\item{I}{Annual heat index, \eqn{I} (\eqn{-}).}

\item{day}{Number of days in a month, \eqn{day} (\eqn{days}).}

\item{a}{An emperical coefficient, assumed to be 1.6.}

\item{b}{An emperical coefficient, assumed to be 6.75e-7.}

\item{c}{An emperical coefficient, assumed to be 7.71e-5.}

\item{d}{An emperical coefficient, assumed to be 1.79e-2.}

\item{e}{An emperical coefficient, assumed to be 0.49.}
}
\description{
\code{E_Thornthwaite} returns the daily evaporation, \eqn{E} (\eqn{mm
day^{-1}}) using temperature data.
}
\details{
\eqn{E} is determined via (Thornthwaite 1948, Rosenberry et al. 2007):

\deqn{E = \bigg(a \bigg(\frac{10 T_{a}}{I} \bigg)^{b I^{3} - c I^{2} + d I +
e} \bigg) \frac{10}{day}}

Thornthwaite CW. 1948. An approach toward a rational classification of
climate. Geographical Review 38 (1): 55. DOI: 10.2307/210739.

Rosenberry DO, Winter TC, Buso DC, Likens GE. 2007. Comparison of 15
evaporation methods applied to a small mountain lake in the northeastern USA.
Journal of Hydrology 340 (3–4): 149–166. DOI: 10.1016/j.jhydrol.2007.03.018.
}
\examples{

}
