% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equil_enrich.R
\name{equil_enrich}
\alias{equil_enrich}
\title{Liquid-vapor equilibrium enrichment factor}
\usage{
equil_enrich(alpha_VL)
}
\arguments{
\item{alpha_VL}{Liquid-vapor equilibrium fractionation factor,
\eqn{\alpha_{V/L}} [\eqn{-}].}
}
\description{
\code{equil_enrich} estimates the isotopic enrichment, \eqn{\epsilon_{V/L}}
(\eqn{\text{\textperthousand}}), given a known liquid-vapor equilibrium
fractionation.
}
\details{
Enrichment is determined via (Horita et al. 2008):

\deqn{\epsilon_{V/L} = (1 - \alpha_{V/L})10^{3}}

Horita J, K Rozanski, S Cohen. 2008. Isotope effects in the evaporation of
water: A status report of the Craig-Gordon model. Isotopes in environmental
and health studies 44 (1): 23–49. DOI: 10.1080/10256010801887174.
}
\examples{

}
