% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equil_frac.R
\name{equil_frac}
\alias{equil_frac}
\title{Liquid-vapor equilibrium fractionation}
\usage{
equil_frac(temperature, element)
}
\arguments{
\item{temperature}{Numeric value of absolute temperature (Kelvin)}

\item{element}{Character indicating "Hydrogen" or "Oxygen"}
}
\value{
Returns the fractionation factor (aka, isotope effect) between liquid
 and vapor, under equilibrium conditions [-].
}
\description{
Estimates the temperature-dependent equilibrium fracitonation of common water
isotopes.
}
\details{
Majoube M. 1971. Oxygen-18 and deuterium fractionation between water and steam
(in French). Journal de Chimie Physique et de Physico-Chimie Biologique 68:
1423–1436
}
\examples{

}
