% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{fraction_to_ratio}
\alias{fraction_to_ratio}
\title{Convert fractional abundance to isotope ratio}
\usage{
fraction_to_ratio(fracAbund)
}
\arguments{
\item{fracAbund}{The fractional abundance of a heavy isotope, \eqn{[-]}.}
}
\value{
A unitless numeric value.
}
\description{
\code{fraction_to_ratio} converts the fractional abundance of a heavy isotope
to its corresponding isotope ratio. For example, the fracitonal abundance of
\eqn{{}^{18}O} is \eqn{{}^{18}F = {}^{18}O / ({}^{16}O + {}^{18}O)}.
}
\details{
\deqn{{}^{A}R = {}^{A}F (1 - {}^{A}F)}

Where A is the heavy isotope of interest, F is the fractional abundance
\eqn{[-]}, and R is the isotope ratio \eqn{[-]}.
}
