% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HH16O_diff.R
\name{HH16O_diff}
\alias{HH16O_diff}
\title{Diffusion of HH16O}
\usage{
HH16O_diff(temperature, Do = 1.635e-08, temp.s = 215.05, gamma = 2.063)
}
\arguments{
\item{temperature}{Vector of air temperatures (C). \eqn{T}}

\item{Do}{Diffussion coefficient (1e-9 m^2 s^-1)}

\item{temp.s}{Standardized temperature (Kelvin)}

\item{gamma}{Fitted exponent [-]}
}
\value{
Returns a temperature-dependent estimate of molecular diffsuion for
  HH16O water (1e-9 m^2 s^-1).
}
\description{
\code{HH16O_diff} implements a temperature-dependent estimate of diffusion
for water - in particular HH16O.
}
\details{
Diffusion of HH16O as determined by Holz et al. (2000):

\deqn{ln(10^{9}D) = a_{1} + a_{2}(1000/T) + a_{3}(1000/T)^{2}}

\tabular{rrrrrr}{
Where:\cr
\deqn{a_{1}} \tab \tab  1.67662250\cr
\deqn{a_{2}} \tab \tab  1.68167989\cr
\deqn{a_{3}} \tab \tab -0.577341011
}

Holz M, SR Heil , A Sacco. 2000. Temperature-dependent self-diffusion
coefficients of water and six selected molecular liquids for calibration in
accurate 1H NMR PFG measurements. Physical Chemistry Chemical Physics 2 (20):
4740 - 4742. DOI: 10.1039/b005319h.
}
\examples{
Temps <- 288:308
HH16O.diffusion <- HH16O_diff(Temps)
ggplot2::qplot(Temps, HH16O.diffusion, geom = "point")
}
