% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lat_heat_vap.R
\name{lat_heat_vap}
\alias{lat_heat_vap}
\title{Latent heat of vaporization}
\usage{
lat_heat_vap(Tl)
}
\arguments{
\item{Tl}{Lake temperature, \eqn{T_{l}} (\eqn{C}).}
}
\description{
\code{lat_heat_vap} returns the latent heat of vaporation
}
