% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{permille_to_decimal}
\alias{permille_to_decimal}
\title{Permille to decimal conversion}
\usage{
permille_to_decimal(delta)
}
\arguments{
\item{delta}{A \eqn{\delta} value in units of permille.}
}
\value{
Numeric value.
}
\description{
As a convenience, \code{permille_to_decimal} converts delta values from units
of permille to decimal values.
}
