% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psychrometric.R
\name{psychrometric}
\alias{psychrometric}
\title{Psychrometric constant}
\usage{
psychrometric(P = 101.3, ca = 0.001, L = 2.47)
}
\arguments{
\item{P}{Air pressure, \eqn{P} (\eqn{kPa}). Assumed to be 101.3, a common
value at sea level.}

\item{ca}{Heat capacity of air, \eqn{c_{a}} (\eqn{MJ kg^{-1} K^{-1}}).
Assumed to be 1e-3.}

\item{L}{Latent heat of vaporization for water, \eqn{\lambda_{v}}
(\eqn{MJ kg^{-1}}). Assumed to be 2.47.}
}
\description{
Determines the psychrometric constant, \eqn{\gamma} (\eqn{kPa K^{-1}}), at a
given pressure.
}
\details{
\eqn{\gamma} is defined as (Dingman 2008):

\deqn{\gamma = \frac{c_{a} P}{0.622 \lambda_{v}}}

Dingman SL. 2008. Physical hydrology. Waveland Press.
}
\examples{

}
