% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{ratio_to_delta}
\alias{ratio_to_delta}
\title{Convert isotope ratio to delta value}
\usage{
ratio_to_delta(ratio, element, standard = "VSMOW", unit)
}
\arguments{
\item{ratio}{Heavy isotope ratio of either hydrogen or oxygen, \eqn{[-]}.}

\item{element}{Character value indicating either "Hydrogen" or "Oxygen".}

\item{standard}{Character value indicating the acronym of the standard to be
used: "VMSOW", "GISP", or "SLAP".}

\item{unit}{Character value explicitly stating if the output value should be
in "decimal" or "permille" units.}
}
\value{
A numeric delta value in either permille or decimal units.
}
\description{
This function, \code{ratio_to_delta} takes the heavy isotope ratio of a known
element, either hydrogen or oxygen, and converts it a delta value (per mille)
for a given standard. Applicable standards include Vienna Standard Mean Ocean
Water (VSMOW), Greenland Ice Sheet Precipitation (GISP), and Standard Light
Antarctic Precipitation (SLAP).
}
\details{
The relationship used to convert a given isotope ratio to delta notation:

\deqn{\delta^{A}X = \Bigg(\frac{{}^{A}R}{{}^{A}R_{Standard}} - 1\Bigg) 1000}

Where \eqn{A} indicates the isotope of element \eqn{X}, \eqn{R} is the ratio
of the sample, and \eqn{R_{Standard}} is the isotope ratio of the standard
being used.

Below are the values of each standard and its isotope ratio used in this
function.

\tabular{lrr}{
Standard \tab 2H/1H \tab 18O/16O\cr
VSMOW \tab 0.00015576 \tab 0.00200520\cr
GISP \tab 0.00012624 \tab 0.00195555\cr
SLAP \tab 0.00008902 \tab 0.00189391 }

VSMOw and SLAP standard information can be found at:
https://nucleus.iaea.org/rpst/documents/VSMOW_SLAP.pdf

GISP standard information:
https://nucleus.iaea.org/rpst/Documents/rs_GISP.pdf
}
