% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{ratio_to_fraction}
\alias{ratio_to_fraction}
\title{Convert isotope ratio to fractional abundance}
\usage{
ratio_to_fraction(ratio)
}
\arguments{
\item{ratio}{The isotope ratio of a heavy istope, \eqn{[-]}.}
}
\value{
A numeric value representing an isotope ratio.
}
\description{
\code{ratio_to_fraction} is the inverse of \code{fraction_to_ratio}. For more
information about this function, see documentation for the latter.
}
\details{
Equation used to convert from an isotope ratio to fractional abundance:

\deqn{{}^{A}F = {}^{A}R(1 + {}^{A}R)}

Where \eqn{A} is the isotope of interest (usually the heavy, or less common
one), \eqn{F} is the fractional abundance between the rare and common
isotope, and \eqn{R} is the isotope ratio.

NOTE: This function assumes only two isotopes are important in the system.
}
