% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sat_vap.R
\name{sat_vap}
\alias{sat_vap}
\title{Saturation vapor pressure of air}
\usage{
sat_vap(temperature)
}
\arguments{
\item{temperature}{Air temperature (C). \eqn{T}}
}
\value{
The result is in kPa.
}
\description{
\code{sat_vap} determines the saturation vapor pressure for a given temperature.
}
\details{
Saturation vapor pressure is determine by (Allen et al. Allen et al. 1998):

\deqn{e_{s} = \frac{0.611 exp(T 17.3)}{T + 237.3}}
}
\examples{

}
