% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVD.R
\name{SVD}
\alias{SVD}
\title{Saturation vapor density}
\usage{
SVD(es, Ta, R = 461.5)
}
\arguments{
\item{es}{Saturation vapor pressure, \eqn{e_{s}} (\eqn{kPa})}

\item{Ta}{Air temperature, \eqn{T_{a}} (\eqn{C}).}

\item{R}{Gas constant for water vapor, \eqn{\rho_{s}} (\eqn{J kg^{-1}
K^{-1}}). Assumed to be 461.5.}
}
\value{

}
\description{
\code{SVD} estimates the saturation vapor density, \eqn{\rho_{s}}, returning
a value in \eqn{g m^{-3}}.
}
\details{
\eqn{\rho_{s}} is estimated by (Manton and Cotton 1977):

\deqn{\rho_{s} = \frac{e_{s}}{R \cdot T_{a}}}

Note that values for \eqn{e_{s}}, \eqn{T_{a}}, and \eqn{\rho_{s}} are
converted to unit consistent values within the function.

Manton M, Cotton W. 1977. Formulation of approximate equations for modeling
moist deep convection on the mesoscale. Atmospheric science paper; no. 266.
Colorado State University, Fort Collins, CO.
}
\examples{

}
