% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atm_vap.R
\name{atm_vap}
\alias{atm_vap}
\title{Atmospheric vapor pressure}
\usage{
atm_vap(RH, e_sat)
}
\arguments{
\item{RH}{Relative humidity, \eqn{RH} [\eqn{-}]}

\item{e_sat}{Saturation vapor pressure, \eqn{e_{s_{Ta}}} (\eqn{kPa}).}
}
\description{
\code{atm_vap} returns the ambient atmospheric vapor pressure, \eqn{e_{a}}
(\eqn{kPa}), given a relative humidity and saturation vapor pressure for a
particular temperature.
}
\details{
Atmospheric vapor pressure (Dingman 2008):

\deqn{e_{a} = RH e_{s_{Ta}}}

Dingman SL. 2008. Physical hydrology. Waveland Press.
}
\examples{

}
