% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equil_frac.R
\name{equil_frac}
\alias{equil_frac}
\title{Liquid-vapor equilibrium fractionation}
\usage{
equil_frac(Ta, element)
}
\arguments{
\item{Ta}{Air temperature, \eqn{T_{a}} (\eqn{C}).}

\item{element}{Character indicating "Hydrogen" or "Oxygen".}
}
\description{
Estimates the temperature-dependent equilibrium fracitonation factor (aka,
isotope effect), \eqn{\alpha_{V/L}}, of common water isotopes (i.e.,
2-hydrogen and 18-oxygen) between the liquid and vapor phases. Units of the
output \eqn{\alpha} value are in \eqn{\text{\textperthousand}}.
}
\details{
Fractionation s determined via (Majoube 1971):

Hydrogen: \eqn{10^{3} ln\alpha_{V/L} = 24.844(10^{6}/(T_{a} + 273.15)^{2}) -
76.248(10^{3}/(T_{a} + 273.15)^{2}) + 52.612}

Oxygen: \eqn{10^{3} ln\alpha_{V/L} = 1.137(10^{6}/(T_{a} + 273.15)^{2}) -
0.4156(10^{3}/(T_{a} + 273.15)) - 2.0667}

Majoube M. 1971. Oxygen-18 and deuterium fractionation between water and
steam (in French). Journal de Chimie Physique et de Physico-Chimie Biologique
68: 1423–1436.
}
\examples{

}
