% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equil_frac.R
\name{equil_frac}
\alias{equil_frac}
\title{Liquid-vapor equilibrium fractionation}
\usage{
equil_frac(temperature, element)
}
\arguments{
\item{temperature}{Numeric value of absolute temperature (Kelvin)}

\item{element}{Character indicating "Hydrogen" or "Oxygen"}
}
\value{
Returns the fractionation factor (aka, isotope effect) between liquid
 and vapor, under equilibrium conditions [-].
}
\description{
Estimates the temperature-dependent equilibrium fracitonation of common water
isotopes.
}
\details{
Fractionation s determined via (Majoube 1971):

Hydrogen: \eqn{10^{3}ln\alpha = 24.844(10^{6}/T_{K}^{2}) -
76.248(10^{3}/T_{K}^{2}) + 52.612}

Oxygen: \eqn{10^{3}ln\alpha = 1.137(10^{6}/T_{K}^{2}) -
0.4156(10^{3}/T_{K}^{2}) - 2.0667}

Majoube M. 1971. Oxygen-18 and deuterium fractionation between water and steam
(in French). Journal de Chimie Physique et de Physico-Chimie Biologique 68:
1423–1436
}
\examples{

}
