% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sat_vap_surf.R
\name{sat_vap_surf}
\alias{sat_vap_surf}
\title{Saturation vapor pressure at the water surface}
\usage{
sat_vap_surf(es, del, Tls, Ta)
}
\arguments{
\item{es}{The saturation vapor pressure of the atmosphere,
\eqn{e_{s}} (\eqn{kPa}).}

\item{del}{Slope of the saturation vapor pressure curve, \eqn{\Delta}
(\eqn{kPa C^{-1}}).}

\item{Tls}{Water temperature at the lake surface, \eqn{T_{ls}} (\eqn{C}).}

\item{Ta}{Temperature of the atmosphere, \eqn{T_{a}} (\eqn{C}).}
}
\description{
\code{sat_vap_surf} returns the estimated saturation vapor pressure,
\eqn{e_{ls}}, of a water body at its surface in \eqn{kPa}.
}
\details{
The saturation vapor pressure of a water body is determined given a known air
temperature, lake surface temperature, the slope of the saturation vapor
pressure curve for a given temperature and saturation vapor pressure for the
atmosphere. The relationship used is (Dingman 2008):

\deqn{e_{ls} = e_{s} + \Delta(T_{ls} - T_{a})}

Dingman SL. 2008. Physical hydrology. Waveland Press.
}
\examples{

}
