% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slp_sat_vap.R
\name{slp_sat_vap}
\alias{slp_sat_vap}
\title{Slope of the saturation vapor pressure curve}
\usage{
slp_sat_vap(temperature, es)
}
\arguments{
\item{temperature}{The air temperature above a body of water (C). \eqn{T}}

\item{es}{Saturation vapor pressure of air (kPa). \eqn{e_{s}}}
}
\value{
Returns the slope of the saturation vapor curve (\eqn{kPa C^{-1}}).
}
\description{
Estimates the slope of the saturation vapor pressure curve, given a specific
temperature and saturation vapor pressure curve.
}
\details{
Slope is estimated via (Allen et al. 1998):

\deqn{\Delta = \frac{4098 e_{s}}{(T + 237.3)^{2}}}

Allen RG, Pereira LS, Raes D, Smith M. 1998. Crop evapotranspiration -
Guidelines for computing crop water requirements - FAO Irrigation and
drainage paper 56. FAO, Rome, 300(9), D05109.
}
\examples{

}
